Add RetireToucanCarbon.sol
This contract provides for the retirement of the underlying TCO2 ERC20 tokens that are held inside a Toucan Protocol Carbon Pool. There is an optional fee that can be set with this contract, and all fees will be bonded into the respective Klima DAO bond repository when the bond size is large enough.

Users can call retireWithPool to use the native pool token that is being redeemed and retired, or call retireWithUSDC/retireWithKLIMA to have those assets swapped through the mapped AMM router for the needed pool tokens prior to redemption/retirement. Any leftover dust from the trade is returned to the user in the form it was received (ex. wsKLIMA is returned as wsKLIMA, even though it had to be unwrapped, unstaked, swapped, then dust re-staked, and re-wrapped).