Incorporate feedback into RetireToucanCarbon.sol Add retirement storage contract.
Feedback from PR reviews incorporated to RetireToucanCarbon.sol.

KlimaCarbonRetirements.sol is a new central storage location to be used for tracking the total carbon retired for a specific address. Some features.
- Updates must be made by a whitelisted helper contract.
- Potential reward claims must be made by a whitelisted contract.

Much of this storage data is duplicative with events emitted, but is being done for the ease of integrating potential future reward and retirement helper contracts.