Ignore less than a kilo retirements
Going forward every retirement is going to be
tracked as part of the NFT certificate contract
so it's easy for users to mint a single certificate
with multiple retirements or any combination of
retirements to NFT certificates.

Currently, the smallest denomination in the NFT
certificate is chosen to be a kilogram so any
smaller denomination will be invalid.