Account for redemption fee in swap
Add the selective redemption fee to the total amount of pool tokens needed in the swap as part of the `_prepareRetireSpecific` function.

Also use the balance of the pool tokens at the end of the `retireCarbonSpecific` function to send to the treasury as a fee to handle any edge cases where multiple TCO2s are redeemed and rounding could cause an issue.