Add C3 Helper Contract
New interface files are present as current UBO/NBO liquidity resides in the Sushiswap Trident AMM platform. The helper performs a two step swap if needed to go from non-Klima tokens to Klima on the legacy UniswapV2 contracts, then swaps from Klima to the pool token in Trident. There is also no `swapForExact` style function within Trident, so the best approximation is made and there will be a slight amount of dust included in the fee sent to the Klima treasury.

The Master aggregator contract is updated to account for the new bridge and the logic needed to direct retirements to that helper.